[김동영] 수강 등록, 수정 추가

+++ on-off/src/main/java/com/greedy/onoff/member/service/MemberService.java
@@ -211,7 +211,7 @@ public class MemberService {
 				String imageName = UUID.randomUUID().toString().replace("-", "");
 				replaceFileName = FileUploadUtils.saveFile(IMAGE_DIR, imageName, memberDto.getMemberImage());
 				memberDto.setMemberImageUrl(replaceFileName);
-				
+				if(oriImage != null) 
 				/* 기존에 저장 된 이미지 삭제*/
 				FileUploadUtils.deleteFile(IMAGE_DIR, oriImage);
